+++ backend/src/services/defaults/index.ts
@@ -1,0 +1,3 @@
+import { GetLandingPage } from "./GetLandingPage";
+
+export default [GetLandingPage];
